Fixed signed and unsigned >=

+++ Seth/Hacks/Animations.cpp
@@ -118,14 +118,14 @@ void Animations::handlePlayers(FrameStage stage) noexcept
 
 Animations::Players Animations::getPlayer(int index) noexcept
 {
-    if (index >= players.size())
+    if (index >= static_cast<int>(players.size()))
         return {};
     return players.at(index);
 }
 
 Animations::Players* Animations::setPlayer(int index) noexcept
 {
-    if (index >= players.size())
+    if (index >= static_cast<int>(players.size()))
         return {};
     return &players.at(index);
 }
